requirements.txt
flask
whisper
ffmpeg-python
pydub
sentence-transformers
nltk
